requirements.txt
dnspython==2.4.2
requests==2.31.0
aiohttp==3.8.4
pytest==7.4.0
rich==13.3.0
